fix: route changes according to module and service renaming.
