+++ apps/nest-datalayer-checker/src/web-agent/action/handlers/handler.module.ts
@@ -1,4 +1,3 @@
-import { ClickStrategy } from './../strategies/click-strategies/utils';
 import { Module } from '@nestjs/common';
 import { ChangeHandler } from './change-handler.service';
 import { ClickHandler } from './click-handler.service';
@@ -7,7 +6,7 @@ import { UtilitiesModule } from '../../utilities/utilities.module';
 import { ChangeStrategyModule } from '../strategies/change-strategies/change-strategy.module';
 import { ClickStrategyModule } from '../strategies/click-strategies/click-strategy.module';
 import { HoverStrategyModule } from '../strategies/hover-strategies/hover-strategy.module';
-import { SharedModule } from '../../../shared/shared.module';
+import { OsModule } from '../../../os/os.module';
 import { ClickStrategyService } from '../strategies/click-strategies/click-strategy.service';
 
 const handlers = [ChangeHandler, ClickHandler, HoverHandler];
@@ -19,7 +18,7 @@ const strategies = [
 ];
 
 @Module({
-  imports: [UtilitiesModule, SharedModule, ...strategies],
+  imports: [UtilitiesModule, OsModule, ...strategies],
   providers: [...handlers, ClickStrategyService],
   exports: [...handlers, ClickStrategyService],
 })
