+++ apps/nest-datalayer-checker/src/web-agent/action/strategies/click-strategies/click-strategy.module.ts
@@ -1,5 +1,5 @@
 import { Module } from '@nestjs/common';
-import { SharedModule } from './../../../../shared/shared.module';
+import { OsModule } from './../../../../os/os.module';
 import { EvaluateClickService } from './evaluate-click.service';
 import { PageClickService } from './page-click.service';
 import { ClickStrategyService } from './click-strategy.service';
@@ -7,7 +7,7 @@ import { ClickStrategyService } from './click-strategy.service';
 const operationStrategies = [EvaluateClickService, PageClickService];
 
 @Module({
-  imports: [SharedModule],
+  imports: [OsModule],
   providers: [...operationStrategies, ClickStrategyService],
   exports: [...operationStrategies, ClickStrategyService],
 })
