+++ apps/nest-datalayer-checker/src/web-agent/web-monitoring/web-monitoring.module.ts
@@ -1,10 +1,10 @@
 import { Module } from '@nestjs/common';
 import { WebMonitoringService } from './web-monitoring.service';
-import { SharedModule } from '../../shared/shared.module';
+import { OsModule } from '../../os/os.module';
 import { RequestModule } from './request/request.module';
 import { DataLayerModule } from './data-layer/data-layer.module';
 @Module({
-  imports: [SharedModule, RequestModule, DataLayerModule],
+  imports: [OsModule, RequestModule, DataLayerModule],
   providers: [WebMonitoringService],
   exports: [WebMonitoringService],
 })
